+++ profiles/ulf/themes/ulf_default/template.php
@@ -8,6 +8,42 @@
  * Implements hook_preprocess_html().
  */
 function ulf_default_preprocess_html(&$variables) {
+  // Set Open Graph tags for nodes.
+  $nodes = $variables['page']['content']['system_main']['nodes'];
+  if (count($nodes) > 0) {
+    $node = reset($nodes);
+
+    $meta_description = array(
+      '#type' => 'html_tag',
+      '#tag' => 'meta',
+      '#attributes' => array(
+        'name' => 'og:title',
+        'content' => $variables['head_title_array']['title'],
+      )
+    );
+    drupal_add_html_head($meta_description, 'meta_title');
+
+    $imagePath = image_style_url('node_display', $node['field_image'][0]['#item']['uri']);
+
+    $meta_description['#attributes']['name'] = 'og:image';
+    $meta_description['#attributes']['content'] = $imagePath;
+    drupal_add_html_head($meta_description, 'meta_image');
+
+    if (isset($node['#node']->field_teaser['und'][0]['value'])) {
+      $meta_description['#attributes']['name'] = 'og:description';
+      $meta_description['#attributes']['content'] = $node['#node']->field_teaser['und'][0]['value'];
+      drupal_add_html_head($meta_description, 'meta_description');
+    }
+
+    $meta_description['#attributes']['name'] = 'og:url';
+    $meta_description['#attributes']['content'] = url(current_path(), array('absolute' => TRUE));
+    drupal_add_html_head($meta_description, 'meta_url');
+
+    $meta_description['#attributes']['name'] = 'og:locale';
+    $meta_description['#attributes']['content'] = 'da_DK';
+    drupal_add_html_head($meta_description, 'meta_locale');
+  }
+
   if (isset($variables['page']['content']['system_main']['field_profile_name']['0']['#markup'])) {
     $variables['head_title'] = $variables['page']['content']['system_main']['field_profile_name']['0']['#markup'] . ' | ' . $variables['head_title_array']['name'];
   }
@@ -601,3 +637,4 @@ function _ulf_default_create_ranges($arr) {
 
   return $ranges;
 }
+